feat: remove extract frontend url util file to middleware
