+++ src/controllers/companyController.ts
@@ -1,16 +1,17 @@
-import { Request, Response } from 'express';
-import AppException from '../exceptions/appException';
 import { HttpStatusCode } from 'axios';
-import { extractCompanySlug } from '../utils/extractCompanySlugFromEmail';
-import { userService } from '../services/userService';
-import { companyService } from '../services/companyService';
+import { Request, Response } from 'express';
 import { Types } from 'mongoose';
-import { membershipService } from '../services/membershipService';
+
 import { ROLE } from '../config';
-import { clearPendingUserData, getPendingUserData } from '../utils/storagePendingUser';
-import { checkVerificationCode } from '../services/auth/registerService';
-import { RegisterUserInput } from './auth/registerController';
+import AppException from '../exceptions/appException';
+import { extractCompanySlug } from '../middleware/extractFrontendUrl';
 import UserModel, { User } from '../models/user';
+import { checkVerificationCode } from '../services/auth/registerService';
+import { companyService } from '../services/companyService';
+import { membershipService } from '../services/membershipService';
+import { userService } from '../services/userService';
+import { clearPendingUserData, getPendingUserData } from '../utils/storagePendingUser';
+import { RegisterUserInput } from './auth/registerController';
 
 export const companyController = {
   createCompany: async (req: Request, res: Response) => {
