+++ src/middleware/validation/adminRegistrationPreCheck.ts
@@ -1,11 +1,12 @@
-import { Request, Response, NextFunction } from 'express';
-import { extractCompanySlug } from '../../utils/extractCompanySlugFromEmail';
-import AppException from '../../exceptions/appException';
 import { HttpStatusCode } from 'axios';
+import { NextFunction, Request, Response } from 'express';
+
+import { RegisterUserInput } from '../../controllers/auth/registerController';
+import AppException from '../../exceptions/appException';
 import CompanyModel from '../../models/company';
 import UserModel from '../../models/user';
 import { getSafePendingUserData, setPendingUserData } from '../../utils/storagePendingUser';
-import { RegisterUserInput } from '../../controllers/auth/registerController';
+import { extractCompanySlug } from '../extractFrontendUrl';
 
 export const validateRegistration = async (req: Request, res: Response, next: NextFunction) => {
   const { email, username } = req.body;
